pyproject.toml
[tool.poetry]
name = "litestar-test-task"
version = "0.1.0"
description = "Test task for LiteStar REST API"
authors = ["Alex"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1"
litestar = { extras = ["standard"], version = "^2" }
litestar-granian = "^0"
litestar-asyncpg = "^0"
advanced-alchemy = "^0.20"
msgspec = "^0.18.6"
pydantic-settings = "^2"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]